Added separate component for the keyboard
