+++ src/routes/App.js
@@ -1,69 +1,19 @@
-import './App.css';
-import React, { useEffect, useState } from 'react';
+// Import dependencies
+import { useState } from 'react';
+
+// Import components
 import Splash from '../components/splash/splash';
 import Metronome from '../components/metronome/metronome';
 import Volume from '../components/volume/volume';
-import useSound from 'use-sound';
-import keys from "../sounds/keys.mp3";
+import Keyboard from '../components/keyboard/keyboard';
 
-const useKeyboardBindings = (play, setKey, splash, keys) => {
-  useEffect(() => {
-    if(!splash) {
-      const keyHandler = e => {
-        if(document.activeElement.id !== 'bpmInput' && keys.includes(e.key)) {
-          play({ id: e.key });
-          setKey(e.key);
-  
-          const element = document.getElementById('key-' + e.key);
-  
-          if(element.classList.contains('keyPressed')) {
-              element.style.animation = 'none';
-              setTimeout(function(){ 
-                  element.style.animation = '';
-              }, 1); 
-          } else {
-            element.classList.add('keyPressed');
-          }
-        }
-      }
-
-      window.addEventListener('keydown', keyHandler);
-  
-      return () => {
-        window.removeEventListener('keydown', keyHandler);
-      };
-    }
-  }, [play, setKey, splash, keys]);
-};
+// Extra files
+import './App.css';
 
 function App() {
   let [splash, setSplash] = useState(true);
   let [currentKey, setKey] = useState('🎹');
   let [volume, setVolume] = useState(50);
-
-  const ordered = ['$', 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 'a', 'b', 'B', 'c', 'C', 'd', 'D', 'e', 'E',
-    'f', 'g', 'G', 'h', 'H', 'i', 'I', 'j', 'J', 'k', 'l', 'L', 'm', '!', '%', '(',
-    '*', '@', '^', 'n', 'o', 'O', 'p', 'P', 'q', 'Q', 'r', 's', 'S', 't', 'T', 'u',
-    'v', 'V', 'w', 'W', 'x', 'y', 'Y', 'z', 'Z'];
-
-  const keyboard = ['1', '!', '2', '@', '3', '4', '$', '5', '%', '6', '^', '7', '8', 
-    '*', '9', '(', '0', 'q', 'Q', 'w', 'W', 'e', 'E', 'r', 't', 'T', 'y', 'Y', 'u',
-    'i', 'I', 'o', 'O', 'p', 'P', 'a', 's', 'S', 'd', 'D', 'f', 'g', 'G', 'h', 'H',
-    'j', 'J', 'k', 'l', 'L', 'z', 'Z', 'x', 'c', 'C', 'v', 'V', 'b', 'B', 'n', 'm'];
-
-  const map = { sprite: {}, volume: volume / 100 };
-
-  ordered.forEach((item, index) => {
-    map.sprite[item] = [(index * 7652), 7648];
-  });
-
-  const [play] = useSound(keys, map);
-  useKeyboardBindings(play, setKey, splash, keyboard);
-
-  const clickPlay = (key) => {
-    play({ id: key });
-    setKey(key);
-  }
 
   // const handleInput = (e) => {
   //   if(e.target.value < 1000) setBpm(e.target.value);
@@ -96,25 +46,9 @@ function App() {
           <Metronome volume={volume}/>
           <Volume setVolume={setVolume} volume={volume}/>
         </div>
+        
+        <Keyboard setKey={setKey} volume={volume} splash={splash}/>
 
-        <div id='keyboard'>
-          {keyboard.map((d, i) =>
-            !isNaN(d) || d !== d.toUpperCase() ?
-              <div id={'section-' + d.toString()} className='section'>
-                <div onMouseDown={() => clickPlay(d.toString())} id={'key-' + d.toString()} className='key'>
-                  <span className='noselect'>{d}</span>
-                </div>
-                {(() => {
-                  if (i + 1 < keyboard.length && keyboard[i + 1] === keyboard[i + 1].toString().toUpperCase()) {
-                    return <div onMouseDown={() => clickPlay(keyboard[i + 1].toString())} id={i + 1} className='black_key'>
-                      <span className='noselect'>{keyboard[i + 1]}</span>
-                    </div>
-                  }
-                })()}
-              </div>
-              : null
-          )}
-        </div>
       </div>
 
       <div id='bottom'>
